feat: Test slot replacement
