+++ src/SlidingPagination.vue
@@ -373,7 +373,7 @@ export default {
           1,
           -1,
           this.ariaPreviousPageLabel,
-          this.slotOrDefault('previous-page', '&laquo;')
+          this.slotOrDefault('previousPage', '&laquo;')
         )
       )
     }
@@ -381,13 +381,13 @@ export default {
     listElements = listElements.concat(this.beginningPages.map((page) => { return this.pageListPage(h, page) }))
 
     if (this.hasBeginningGap) {
-      listElements.push(this.gap(h, this.slotOrDefault('gap-left', '&hellip;')))
+      listElements.push(this.gap(h, this.slotOrDefault('gapLeft', '&hellip;')))
     }
 
     listElements = listElements.concat(this.slidingWindowPages.map((page) => { return this.pageListPage(h, page) }))
 
     if (this.hasEndingGap) {
-      listElements.push(this.gap(h, this.slotOrDefault('gap-right', '&hellip;')))
+      listElements.push(this.gap(h, this.slotOrDefault('gapRight', '&hellip;')))
     }
 
     listElements = listElements.concat(this.endingPages.map((page) => { return this.pageListPage(h, page) }))
@@ -399,7 +399,7 @@ export default {
           this.total,
           1,
           this.ariaNextPageLabel,
-          this.slotOrDefault('next-page', '&raquo;')
+          this.slotOrDefault('nextPage', '&raquo;')
         )
       )
     }
